requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.4
SQLAlchemy==2.0.18
psycopg2-binary==2.9.6
opencv-python==4.8.0.76
numpy==1.24.3
matplotlib==3.7.1
python-dotenv==1.0.0  # For environment variables management
